UI files: use notr="true" instead of old comment="KDE::DoNotExtract"
Usage of "KDE::DoNotExtract" predates the notr attribute and still results
in uic generating code trying to do a (bogus) lookup for the very string.

With notr attribute instead the generated code will only set the given
string as-is. While the given string is just a sample text to help when
working on the UI file and ideally would not appear at all in the final
product, this here at least fixes the also unneeded translation call.