When saving the game, use the proper extension "pgn" in case no extension was provided in the file dialog. The currently used extension and the code in this area is a copy&paste bug from LabPlot's code long time and we correct it now.
There is another problem that needs to be addressed here. Namely, no extension added on default
n the file dialog on non-Plasma desktops (reproduced the problem on IceWM, also reported in BUG: 503374).
This is something that needs to be checked in KF independently of this commit which we
need anyway also for cases where the option "automatically select the extension" was
not checked in the dialog and we need to ensure the proper extension in the code.

BUG: 502963
FIXED-IN: 25.04.2